third_party/WebKit/LayoutTests/presentation/presentation-navigation-multipleurls.html

<!DOCTYPE html>
<html>
<body>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>

// TODO(crbug.com/680298): use iframe to do presentation navigation layout tests
if (location.href.endsWith('#after-reload')) {
  test(function() {
    assert_equals(navigator.presentation.defaultRequest, null);
  }, "Test that navigator.defaultRequest with multiple URLs isn't kept alive after reload.")
} else {
  navigator.presentation.defaultRequest = new PresentationRequest(["https://example.org", "cast://google.com/app_id=deadbeef"]);

  location.href += '#after-reload';
  location.reload();
}

</script>
</body>
</html>
